feature: Hanlde smaller screen sizes and clean up code.
