+++ pages/stands/standsData.tsx
@@ -1,0 +1,138 @@
+export const standsPlacement = [
+  { x: 24.0, y: 54.4 }, //1
+  { x: 22.15, y: 51.3 },//2
+  { x: 20.3, y: 51.3 }, //3
+  { x: 18.4, y: 53.4 }, //4
+  { x: 18.4, y: 55.8 }, //5
+  { x: 15.95, y: 56.9 },//6
+  { x: 13.5, y: 56.9 }, //7
+  { x: 12.0, y: 51.2 }, //8
+  { x: 14.5, y: 50.9 }, //9
+  { x: 16.8, y: 49.7 }, //10
+  { x: 22.0, y: 46.8 }, //11
+  { x: 23.3, y: 44.4 }, //12
+  { x: 25.8, y: 40.0 }, //13
+  { x: 27.4, y: 37.1 }, //14
+  { x: 29.9, y: 35.6 }, //15
+  { x: 29.9, y: 41.9 }, //16
+  { x: 29.9, y: 45.2 }, //17
+  { x: 28.4, y: 46.7 }, //18
+  { x: 26.7, y: 49.2 }, //19
+  { x: 26.7, y: 51.8 }, //20
+  { x: 26.7, y: 54.4 }, //21
+  { x: 47.4, y: 59.2 }, //22
+  { x: 49.7, y: 59.2 }, //23
+  { x: 52.0, y: 59.2 }, //24
+  { x: 54.8, y: 51.0 }, //25
+  { x: 54.8, y: 48.4 }, //26
+  { x: 56.9, y: 48.0 }, //27
+  { x: 58.8, y: 48.0 }, //28
+  { x: 60.5, y: 48.5 }, //29
+  { x: 60.5, y: 51.1 }, //30
+  { x: 72.0, y: 56.9 }, //31
+  { x: 74.4, y: 56.9 }, //32
+  { x: 75.3, y: 64.8 }, //33
+  { x: 74.2, y: 71.2 }, //34
+  { x: 72.2, y: 71.2 }, //35
+  { x: 70.0, y: 71.2 }, //36
+  { x: 69.7, y: 67.0 }, //37
+  { x: 69.7, y: 64.2 }, //38
+  { x: 69.7, y: 61.3 }, //39
+  { x: 66.4, y: 61.3 }, //40
+  { x: 66.4, y: 64.2 }, //41
+  { x: 66.4, y: 67.0 }, //42
+];
+
+export const companiesDay1TopRight: string[] = [
+  'Computas',
+  'Norkart',
+  'Netcompany',
+  'Sykehuspartner HF',
+  'Thales Norway',
+  'Hemit HF',
+  'Kantega',
+  'Netlight',
+  'Gjensidige',
+  'Itera',
+  'Equinor',
+  'NAV IT',
+  'Bekk',
+  'Bearingpoint',
+  'Sikt',
+  'Geomatikk',
+  'Applica Consulting',
+  'Brønnøysundregistrene',
+  'Norconsult Digital',
+  'Vespa.ai',
+  'Bouvet',
+];
+
+export const companiesDay1BottomLeft: string[] = [
+  'Skatteetaten',
+  'Pexip',
+  'Knowit',
+  'Accenture',
+  'NOVA Consulting',
+  'Gintel',
+  'SKIPPED', //Added so that we can skip certain unused stands
+  'Elliptic Labs',
+  'BCG Platinion',
+  'Statens Vegvesen',
+  'SKIPPED', //Added so that we can skip certain unused stands
+  'Atea',
+  'Teledyne',
+  'Capgemini',
+  'H5P Group',
+  'Domstolsadministrasjonen',
+  'Data Nova',
+  'Fremtind',
+  'GE Vernova',
+];
+
+export const companiesDay2TopRight: string[] = [
+  'Computas',
+  'Tietoevry',
+  'Netcompany',
+  'Sticos',
+  'Forte Digital',
+  'Jotron',
+  'FFI',
+  'Mnemonic',
+  'NSM',
+  'Systek',
+  'Equinor',
+  'Safetec',
+  'Frend',
+  'Appfarm',
+  'NorgesGruppen',
+  'twoday',
+  'Clave',
+  'Holte Consulting',
+  'Intility',
+  'TV2',
+  'Bouvet',
+];
+
+export const companiesDay2BottomLeft: string[] = [
+  'Sopra Steria',
+  'Norsk Tipping',
+  'Arm',
+  '24SevenOffice',
+  'DNB',
+  'EY',
+  'Autodesk',
+  'UDI',
+  'Genus',
+  'PwC',
+  'Visma',
+  'Konsberg Gruppen',
+  'Sparebank 1 Utv',
+  'Sparebank 1 SMN',
+  'DNV',
+  'KPMG',
+  'Norsk Helsenett',
+  'Optio Incentives',
+  'GE Vernova',
+  'Å Energi',
+  'H5P Group',
+]
